+++ internal/http/handler/balance/balance.go
@@ -18,8 +18,8 @@ type Balance struct {
 var _ router.BalanceHandler = (*Balance)(nil)
 
 type Request struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
 
 func New(
